use usize instead of u32, add comments correct solutions

+++ src/dec03.rs
@@ -1,7 +1,9 @@
 use std::fs::read_to_string;
 use regex::Regex;
 
-pub fn part1() -> u32 {
+
+/// correct answer is 183788984
+pub fn part1() -> usize {
     let memory = read_to_string("src/in/dec03.in").unwrap();
 
     let outer_re = Regex::new(r"mul\((\d{1,3}),(\d{1,3})\)").unwrap();
@@ -20,7 +22,9 @@ pub fn part1() -> u32 {
     return sum.try_into().unwrap();
 }
 
-pub fn part2() -> u32 {
+
+/// correct answer is 62098619
+pub fn part2() -> usize {
     let memory = read_to_string("src/in/dec03.in").unwrap();
     let outer_re = Regex::new(r"mul\((\d{1,3}),(\d{1,3})\)|do\(\)|don't\(\)").unwrap();
     let inner_re = Regex::new(r"\d{1,3}").unwrap();
